+++ test/Outkeep.Core.Tests/TcpHelperTests.cs
@@ -1,11 +1,12 @@
 ﻿using Moq;
-using Outkeep.Hosting.Properties;
+using Outkeep.Core.Properties;
+using Outkeep.Core.Tcp;
 using System;
 using System.Net;
 using System.Net.Sockets;
 using Xunit;
 
-namespace Outkeep.Hosting.Tests
+namespace Outkeep.Core.Tests
 {
     public class TcpHelperTests
     {
@@ -73,7 +74,7 @@ namespace Outkeep.Hosting.Tests
 
             // act and assert
             var error = Assert.Throws<InvalidOperationException>(() => helper.GetFreePort(1, 3));
-            Assert.Equal(Resources.ThereAreNoFreePortsWithinTheInputRange, error.Message);
+            Assert.Equal(Resources.Exception_ThereAreNoFreePortsWithinTheInputRange, error.Message);
         }
 
         [Fact]
@@ -103,7 +104,7 @@ namespace Outkeep.Hosting.Tests
 
             // act and assert
             var error = Assert.Throws<InvalidOperationException>(() => helper.GetFreeDynamicPort());
-            Assert.Equal(Resources.ExceptionUnexpectedEndpointType, error.Message);
+            Assert.Equal(Resources.Exception_UnexpectedEndpointType, error.Message);
         }
 
         [Fact]
@@ -117,7 +118,7 @@ namespace Outkeep.Hosting.Tests
 
             // act and assert
             var error = Assert.Throws<InvalidOperationException>(() => helper.GetFreeDynamicPort());
-            Assert.Equal(Resources.ExceptionCouldNotOpenADynamicPortForExclusiveUse, error.Message);
+            Assert.Equal(Resources.Exception_CouldNotOpenADynamicPortForExclusiveUse, error.Message);
             Assert.Same(ex, error.InnerException);
         }
