+++ test/Outkeep.Hosting.Tests/TcpListenerWrapperTests.cs
@@ -1,4 +1,5 @@
-﻿using System.Net;
+﻿using Outkeep.Core.Tcp;
+using System.Net;
 using System.Net.Sockets;
 using Xunit;
